refactor(dist): implement helper functions for processing assigned LNodes
